Rewriting sliding future prediction strategy
Reviewing and rewriting much simpler code for prediction of future behavior of stocks. 
Now possible behaviors (classes) are down, sideways and up. Classes are defined over a serie of first derivative of price divided by price that is equivalent to percetange of price change in the current time frame (minutes).