Starting background python predictor
The background worker will keep reading a *.mql5 binary file with enough samples
createad by the metatrader5 expert advisor.
If the binary file is new compared with the previous.
It will train the torch NN classifier and if accuracy and others
criterias are reached it will create a prediction file with time and direction
that the expert advisor will read.